move get_proj_mat_by_coord_type to utils
