+++ mmdet3d/core/bbox/structures/__init__.py
@@ -5,11 +5,12 @@ from .coord_3d_mode import Coord3DMode
 from .depth_box3d import DepthInstance3DBoxes
 from .lidar_box3d import LiDARInstance3DBoxes
 from .utils import (get_box_type, limit_period, mono_cam_box2vis,
-                    points_cam2img, rotation_3d_in_axis, xywhr2xyxyr)
+                    points_cam2img, rotation_3d_in_axis, xywhr2xyxyr,
+                    get_proj_mat_by_coord_type)
 
 __all__ = [
     'Box3DMode', 'BaseInstance3DBoxes', 'LiDARInstance3DBoxes',
     'CameraInstance3DBoxes', 'DepthInstance3DBoxes', 'xywhr2xyxyr',
     'get_box_type', 'rotation_3d_in_axis', 'limit_period', 'points_cam2img',
-    'Coord3DMode', 'mono_cam_box2vis'
+    'Coord3DMode', 'mono_cam_box2vis', 'get_proj_mat_by_coord_type'
 ]
