+++ mmdet3d/models/fusion_layers/point_fusion.py
@@ -4,7 +4,8 @@ from mmcv.runner import BaseModule
 from torch import nn as nn
 from torch.nn import functional as F
 
-from mmdet3d.core.bbox.structures import points_cam2img
+from mmdet3d.core.bbox.structures import (get_proj_mat_by_coord_type,
+                                          points_cam2img)
 from ..builder import FUSION_LAYERS
 from . import apply_3d_transformation
 
@@ -87,22 +88,6 @@ def point_sample(img_meta,
         align_corners=align_corners)  # 1xCx1xN feats
 
     return point_features.squeeze().t()
-
-
-# TODO: 1. move it to another file. This also should be used
-#          to unify draw_*_bbox3d_on_img and in ImVoxelNet.
-#       2. update when refactor cam_intrinsic to cam2img.
-#       3. update when refactor coord_type and box_type to
-#          the same case.
-def get_proj_mat_by_coord_type(img_meta, coord_type):
-    coord_type = coord_type.upper()
-    mapping = {
-        'LIDAR': 'lidar2img',
-        'DEPTH': 'depth2img',
-        'CAMERA': 'cam_intrinsic'
-    }
-    assert coord_type in mapping.keys()
-    return img_meta[mapping[coord_type]]
 
 
 @FUSION_LAYERS.register_module()
